requirements.txt
annotated-types==0.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
ChatterBot2==1.1.0a7
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mathparse==0.1.2
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.4
packaging==24.0
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.8.2
pytz==2024.1
regex==2024.5.15
requests==2.32.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.3.24
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
